requirements.txt
uvicorn[standard]
sse-starlette
starlette
websockets
torch
transformers
diffusers
pillow
huggingface_hub
accelerate
numpy==2.0.0
sentencepiece
RealESRGAN
celery 
redis 
fastapi[all]
flower
pytest
httpx
Authlib
sqlalchemy 
psycopg2-binary 
alembic
sqlalchemy-utils
python-jose 
pydantic
pathlib
google-auth
python-json-logger 